**GIANT HISTORY REWRITE**
Prior history was derived from commit 8fd85647e6955bb307d64444efe5426664085d92
of the wwt-web-client repository using the following command:

```
git filter-branch -f \
  --subdirectory-filter webclient \
  --index-filter 'git rm --cached --ignore-unmatch -r \
      searchdataraw.js searchdata.min.js wwtwebclient.js wwtwebclient.min.js \
      sdk packages ext/*.min.js ext/*.min.css m51.txt css/webclient.css* \
      css/webclient.min.css clientbin WWTSL.xap Bin .vs obj Images Scripts \
      WebClient poly.html fblogin.js Chandra.wtml ESO.wtml Hubble.wtml Spitzer.wtml \
      Wise.wtml adsass.wtml gettours_webclient.xml login.aspx login.aspx.cs \
      Global.asax Properties Web.config ensurenpm.cmd' \
  -- HEAD
```

This basically takes the history of that repository, pretends that
`webclient/` was the root of it, and rewrites the history to only include
files from that directory, deleting a variety of generated and/or static
content files that we'll try to avoid tracking in Git.

Based on the experience of the StarHunt project, it looks like we can rebuild
this project to be independent of IIS and Visual Studio, massive lowering the
barrier to entry for our target audience of potential developers.