controllers/MainController.js: fix the cross-fader
It turns out I was a bit too quick in my testing when I put mobile support
back. Putting the desktop HTML code inside a `<div ng-if="...">` turns out to
mean that when the MainController initializer is being run, *neither* the
mobile nor the desktop DOM elements exist yet. So, initialization code that
depends on things that live inside these elements must be queued up to run
later.