index.html: fix desktop imagery chooser
Yikes: the other consequence of adding the mobile support back using AngularJS
ng-if directives. It turns out that ng-if creates an additional layer of
scope, so that if you have an element using `ng-model` inside an `ng-if` that
wasn't there before, changes in the view modify a variable in a different
scope than they did previously. For almost everything here that's OK because
there are ng-controller items that re-anchor the appropriate scope, but not
these little context panel items. It's not the most beautiful, but adding a
$parent scope reference gets things working again.