Reorganize the mobile nav bar
Now there are three buttons: hamburger menu, sharing, and settings. The
hamburger menu gives a subset of the links from the "AAS WorldWide Telescope"
menu of the desktop version; the links from the existing settings menu are
removed.

Both the hamburger menu and the settings are implemented as collapsible navs,
so we need to edit the hideMenu() logic to prevent both menus from being open
at the same time.

The positioning of the search submenu of the Settings nav was a bit weird, I
think accidentally, so I've reworked how that's done so that it's consistent
with the other popups.

I've also made the buttons a bit larger overall (@font-size-large to
@font-size-h3), so a few other CSS settings changed too.