index.html: use ng-show to switch from skyball to planetary scaler
The ng-if directive creates and destroys DOM elements, and so the bindings
that we set up programmatically are lost when cycling between different modes,
resulting in a broken skyball and inability to change planet scale. Recently
reported by Pat, but we didn't change anything here *and* we haven't rolled
out the new webclient implementation, so I'm really not sure what's going on.
Seems improbable that no one noticed before?

Closes #285.