controllers/MainController.js: hardcode that Planet mode defauls to looking at Mars
Part of addressing #293. The ordering of the items in the Planet "Imagery"
dropdown is set by the `builtin-image-sets.wtml` file loaded by the web
engine. We can now conveniently update that (and it's separate from the one
used by the Windows client), but we do want the client to still default to
looking at Mars rather than the Sun when you switch to Planet mode. With this
in place the behavior will be the same.