factories/AutohidePanels.js: hack to apply miscellaneous settings on startup
The "Settings" panel contains a few miscellaneous application settings.
Unfortunateley those settings are only defined in the AngularJS Controller for
that panel's UI, which means that the settings aren't initialized until the UI
is opened. This should be fix for real by turning those settings into a
different kind of AngularJS state object, but in the meantime, we can (ab)use
the AutohidePanels factory, which is initialized earlier, to apply the saved
settings.

We also remove the "Crosshairs" checkbox from the Layer Manager, because that
tracks how the Windows UI works and the two separate checkboxes for the same
setting were getting out of sync.