Get the thumbnail property panel to stay onscreen
If you tried to open the "Properties" panel for a thumbnail in the nearby
objects panel, it would land below the screen. Besides there being CSS
hardcoding the popover position, the whole popover content was encased in an
"ng-if" directive that meant that when the popover position was being
calculated, the apparent size of the popover was tiny and so the
bounds-checking didn't do much of anything. As far as I can tell, we can skip
the ng-if with proper setup of the scope of the popover in its creation.

Closes #91.