controllers/MainController.js: do not auto fullscreen when playing tours
To some extent, this is a UX judgment call. I do not like the auto-fullscreen
behavior, and I think it violates most people's expectations of how web apps
should behave.

It is also true that browsers are adding restrictions that make it so that
webpages can only activate fullscreen mode in response to a user UI
interaction. The way things are currently wired up, that requirement isn't met
consistently when starting to play a tour.

We should add a nice Go Fullscreen button as seen in YouTube or other kinds of
web players. This change doesn't add that.