Don't set the "hidden" DOM property of the finder scope element.
On recent Chrome, turning "hidden" on and off seems to somehow interfere with
event handling, such that the buttons on the finder scope weren't working
anymore [1]. Despite lots of research I can't figure out what exactly is causing
the problem, but it's very likely a new behavior since this code must have
been tested on Chrome historically.

The jQuery fadeIn/fadeOut functions end up altering the element CSS to use
"display: none", which as far as I can tell is an equivalent flag for our
purposes. So if we don't mess with "hidden" at all and just pay attention to
the "display" state, things seem to work again.

Fixes #312.

[1]: For posterity: in my testing, click events seemed to be going through,
but they would be targeting the toplevel `body` element, not the `a.btn`
elements, breaking the processing used by jQuery/Angular. Bizarre.