controllers/MainController.js: attempt to fix #316
Currently, on a Mac, if you try to roll the view with a control-drag, it
also pops up the finder scope. Here we try to prevent this in a somewhat
hacky way. It might be better to preserve the Mac semantics that
control-click = context menu, but that would require selecting a
different mouse gesture for rolling the camera.

Closes #316.