Update hosted engine URLs
Have the production webclient point to `engine/7/`, where we've started
publishing (compared to using `$major.$minor` for the version before).

Have the testing webclient point to `engine/latest/`. I had some idea
about how it was better for the testing webclient to still refer to
published versions of the engine, but it's been a bit annoying I think I
was just wrong before.