ci/azure-deployment.yml: update CDN purge command
I need to update the command here to track some Azure toolkit change. I
ran into the same issue in the `wwt-webgl-engine` repo, so I'm confident
in this fix. But also there's no real way to test it besides merging and
running the deployment logic, so might as well push straight to
`master`.